WeatherApiService/src/test/java/com/skyapi/weatherforecast/common/location/LocationRepositoryTests.java
package com.skyapi.weatherforecast.common.location;

import com.skyapi.weatherforecast.common.Location;
import com.skyapi.weatherforecast.common.location.LocationRepository;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase.Replace;
import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
import org.springframework.test.annotation.Rollback;

import static org.assertj.core.api.Assertions.assertThat;

@DataJpaTest
@AutoConfigureTestDatabase(replace = Replace.NONE)
@Rollback(false)
public class LocationRepositoryTests {
    @Autowired
    private LocationRepository repository;

    @Test
    public void testGetNotFound(){
        String code  = "ABCD";
        Location location = repository.findByCode(code);

        assertThat(location).isNull();
    }
}
